fix(cli): advertise ClinePass in ACP
